move delimiter to common
