+++ demehin.maxim/T2/data_struct.cpp
@@ -5,27 +5,11 @@
 #include <iomanip>
 #include <set>
 #include "scope_guard.hpp"
+#include <delimiter.hpp>
 
 namespace
 {
   using namespace demehin::io;
-}
-
-std::istream& demehin::io::operator>>(std::istream& in, DelimiterIO&& dest)
-{
-  std::istream::sentry sentry(in);
-  if (!sentry)
-  {
-    return in;
-  }
-
-  char c = '0';
-  in >> c;
-  if (in && (std::tolower(c) != dest.exp))
-  {
-    in.setstate(std::ios::failbit);
-  }
-  return in;
 }
 
 std::istream& demehin::io::operator>>(std::istream& in, DoubleIO&& dest)
