+++ demehin.maxim/common/delimiter.cpp
@@ -1,0 +1,23 @@
+#include "delimiter.hpp"
+
+namespace
+{
+  using namespace demehin::io;
+}
+
+std::istream& demehin::io::operator>>(std::istream& in, DelimiterIO&& dest)
+{
+  std::istream::sentry sentry(in);
+  if (!sentry)
+  {
+    return in;
+  }
+
+  char c = '0';
+  in >> c;
+  if (in && (std::tolower(c) != dest.exp))
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
