+++ demehin.maxim/common/delimiter.hpp
@@ -1,0 +1,16 @@
+#ifndef DELIMITER_HPP
+#define DELIMITER_HPP
+#include <istream>
+
+namespace demehin::io
+{
+  struct DelimiterIO
+  {
+    char exp;
+  };
+
+  std::istream& operator>>(std::istream& in, DelimiterIO&& dest);
+}
+
+
+#endif
